requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.40
plotly==5.22.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
